tambahan crud admin

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\EventController;
 use App\Http\Controllers\PostController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\TagController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -29,61 +30,44 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-// Route::get('user', function () {
-//     return '<h1>Hi Siswa</h1>';
-// })->middleware(['auth', 'verified', 'role:user|admin']); //2 user
-
 // Dashboard
-Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
+Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard')->middleware(['auth', 'verified', 'role:admin']);
 
-//test template login
-Route::get('/testlogin', function () {
-    return view('auth.logincopy');
-})->name('test.login');
-
-//test template login
-Route::get('/testregister', function () {
-    return view('auth.registercopy');
-})->name('test.register');
-
-//User
-Route::get('/user', function () {
-    return view('admin.user.index');
-})->name('user.index');
-Route::get('/user/create', function () {
-    return view('admin.user.create');
-})->name('user.create');
+//User Admin
+Route::get('/user', [UserController::class, 'index'])->name('user')->middleware(['auth', 'verified', 'role:admin']);
+Route::delete('/user/delete/{id}', [UserController::class, 'destroy'])->name('user.delete')->middleware(['auth', 'verified', 'role:admin']);
 
 //Post
-Route::get('/post', [PostController::class, 'index'])->name('post');
-Route::get('/post/create', [PostController::class, 'create'])->name('post.create');
-Route::post('/post/create', [PostController::class, 'store'])->name('post.store');
-Route::get('/post/edit/{id}', [PostController::class, 'edit'])->name('post.edit');
-Route::put('/post/update/{post}', [PostController::class, 'update'])->name('post.update');
-Route::delete('/post/delete/{id}', [PostController::class, 'destroy'])->name('post.delete');
+Route::get('/post', [PostController::class, 'index'])->name('post')->middleware(['auth', 'verified', 'role:admin']);
+Route::get('/post/create', [PostController::class, 'create'])->name('post.create')->middleware(['auth', 'verified', 'role:admin']);
+Route::post('/post/create', [PostController::class, 'store'])->name('post.store')->middleware(['auth', 'verified', 'role:admin']);
+Route::get('/post/edit/{id}', [PostController::class, 'edit'])->name('post.edit')->middleware(['auth', 'verified', 'role:admin']);
+Route::put('/post/update/{post}', [PostController::class, 'update'])->name('post.update')->middleware(['auth', 'verified', 'role:admin']);
+Route::delete('/post/delete/{id}', [PostController::class, 'destroy'])->name('post.delete')->middleware(['auth', 'verified', 'role:admin']);
 
 // Tag
-Route::get('/tag', [TagController::class, 'index'])->name('tag');
-Route::get('/tag/create', [TagController::class, 'create'])->name('tag.create');
-Route::post('/tag/create', [TagController::class, 'store'])->name('tag.store');
-Route::get('/tag/edit/{id}', [TagController::class, 'edit'])->name('tag.edit');
-Route::put('/tag/update/{id}', [TagController::class, 'update'])->name('tag.update');
-Route::delete('/tag/delete/{id}', [TagController::class, 'destroy'])->name('tag.delete');
+Route::get('/tag', [TagController::class, 'index'])->name('tag')->middleware(['auth', 'verified', 'role:admin']);
+Route::get('/tag/create', [TagController::class, 'create'])->name('tag.create')->middleware(['auth', 'verified', 'role:admin']);
+Route::post('/tag/create', [TagController::class, 'store'])->name('tag.store')->middleware(['auth', 'verified', 'role:admin']);
+Route::get('/tag/edit/{id}', [TagController::class, 'edit'])->name('tag.edit')->middleware(['auth', 'verified', 'role:admin']);
+Route::put('/tag/update/{id}', [TagController::class, 'update'])->name('tag.update')->middleware(['auth', 'verified', 'role:admin']);
+Route::delete('/tag/delete/{id}', [TagController::class, 'destroy'])->name('tag.delete')->middleware(['auth', 'verified', 'role:admin']);
 
 // Category
-Route::get('/kategori', [CategoryController::class, 'index'])->name('kategori');
-Route::get('/kategori/create', [CategoryController::class, 'create'])->name('kategori.create');
-Route::post('/kategori/create', [CategoryController::class, 'store'])->name('kategori.store');
-Route::get('/kategori/edit/{id}', [CategoryController::class, 'edit'])->name('kategori.edit');
-Route::put('/kategori/update/{id}', [CategoryController::class, 'update'])->name('kategori.update');
-Route::delete('/kategori/delete/{id}', [CategoryController::class, 'destroy'])->name('kategori.delete');
+Route::get('/kategori', [CategoryController::class, 'index'])->name('kategori')->middleware(['auth', 'verified', 'role:admin']);
+Route::get('/kategori/create', [CategoryController::class, 'create'])->name('kategori.create')->middleware(['auth', 'verified', 'role:admin']);
+Route::post('/kategori/create', [CategoryController::class, 'store'])->name('kategori.store')->middleware(['auth', 'verified', 'role:admin']);
+Route::get('/kategori/edit/{id}', [CategoryController::class, 'edit'])->name('kategori.edit')->middleware(['auth', 'verified', 'role:admin']);
+Route::put('/kategori/update/{id}', [CategoryController::class, 'update'])->name('kategori.update')->middleware(['auth', 'verified', 'role:admin']);
+Route::delete('/kategori/delete/{id}', [CategoryController::class, 'destroy'])->name('kategori.delete')->middleware(['auth', 'verified', 'role:admin']);
 
 // Event
-Route::get('/agenda', [EventController::class, 'index'])->name('agenda');
-Route::get('/agenda/create', [EventController::class, 'create'])->name('agenda.create');
-Route::post('/agenda/create', [EventController::class, 'store'])->name('agenda.store');
-Route::get('/agenda/edit/{id}', [EventController::class, 'edit'])->name('agenda.edit');
-Route::put('/agenda/update/{id}', [EventController::class, 'update'])->name('agenda.update');
-Route::delete('/agenda/delete/{id}', [EventController::class, 'destroy'])->name('agenda.delete');
+Route::get('/agenda', [EventController::class, 'index'])->name('agenda')->middleware(['auth', 'verified', 'role:admin']);
+Route::get('/agenda/create', [EventController::class, 'create'])->name('agenda.create')->middleware(['auth', 'verified', 'role:admin']);
+Route::post('/agenda/create', [EventController::class, 'store'])->name('agenda.store')->middleware(['auth', 'verified', 'role:admin']);
+Route::get('/agenda/edit/{id}', [EventController::class, 'edit'])->name('agenda.edit')->middleware(['auth', 'verified', 'role:admin']);
+Route::put('/agenda/update/{id}', [EventController::class, 'update'])->name('agenda.update')->middleware(['auth', 'verified', 'role:admin']);
+Route::delete('/agenda/delete/{id}', [EventController::class, 'destroy'])->name('agenda.delete')->middleware(['auth', 'verified', 'role:admin']);
+
 
 require __DIR__.'/auth.php';
